+++ src/mnelab/io/writers.py
@@ -4,9 +4,7 @@
 
 from pathlib import Path
 
-import mne
-import numpy as np
-from numpy.rec import fromarrays
+from numpy.core.records import fromarrays
 from scipy.io import savemat
 
 
@@ -57,33 +55,9 @@ def write_edf(fname, raw):
     raw.export(fname)
 
 
-def write_bv(fname, raw, events=None):
+def write_bv(fname, raw):
     """Export data to BrainVision EEG/VHDR/VMRK file (requires pybv)."""
-    import pybv
-
-    name, _ = Path(fname).stem, "".join(Path(fname).suffixes)
-    parent = Path(fname).parent
-    data = raw.get_data()
-    fs = raw.info["sfreq"]
-    ch_names = raw.info["ch_names"]
-    if events is None:
-        if raw.annotations:
-            events = mne.events_from_annotations(raw, regexp=None)[0]
-            dur = raw.annotations.duration * fs
-            events = np.column_stack([events[:, [0, 2]], dur.astype(int)])
-    else:
-        events = events[:, [0, 2]]
-    units = [ch["unit"] for ch in raw.info["chs"]]
-    units = ["µV" if unit == 107 else "AU" for unit in units]
-    pybv.write_brainvision(
-        data=data,
-        sfreq=fs,
-        ch_names=ch_names,
-        fname_base=name,
-        folder_out=parent,
-        events=events,
-        unit=units,
-    )
+    raw.export(fname=Path(fname).with_suffix(".vhdr"))
 
 
 # this dict contains each supported file extension as a key; the corresponding value is